# LunchFinder/Api/swagger.yaml
x-generator: NSwag v14.0.7.0 (NJsonSchema v11.0.0.0 (Newtonsoft.Json v13.0.0.0))
openapi: 3.0.0
info:
  title: My Title
  version: 1.0.0
paths:
  /api/Tags:
    get:
      tags:
      - Tag
      operationId: GetAllTags
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
    post:
      tags:
      - Tag
      operationId: CreateTag
      requestBody:
        x-name: tag
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
        required: true
        x-position: 1
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
  /api/Tags/{id}:
    get:
      tags:
      - Tag
      operationId: GetTagById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
        x-position: 1
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        404:
          description: ''
    put:
      tags:
      - Tag
      operationId: UpdateTag
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
        x-position: 1
      requestBody:
        x-name: tag
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
        required: true
        x-position: 2
      responses:
        200:
          description: ''
        404:
          description: ''
    delete:
      tags:
      - Tag
      operationId: DeleteTag
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
        x-position: 1
      responses:
        200:
          description: ''
        404:
          description: ''
  /api/Place:
    get:
      tags:
      - Place
      operationId: GetAllPlaces
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Place'
    post:
      tags:
      - Place
      operationId: CreatePlace
      requestBody:
        x-name: place
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Place'
        required: true
        x-position: 1
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
  /api/Place/{id}:
    get:
      tags:
      - Place
      operationId: GetPlaceById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
        x-position: 1
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
        404:
          description: ''
    put:
      tags:
      - Place
      operationId: UpdatePlace
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
        x-position: 1
      requestBody:
        x-name: place
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Place'
        required: true
        x-position: 2
      responses:
        200:
          description: ''
        404:
          description: ''
    delete:
      tags:
      - Place
      operationId: DeletePlace
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
        x-position: 1
      responses:
        200:
          description: ''
        404:
          description: ''
components:
  schemas:
    Tag:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
    Place:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        address:
          $ref: '#/components/schemas/Address'
        contactInformation:
          type: array
          items:
            $ref: '#/components/schemas/ContactInformation'
    Address:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int32
        street:
          type: string
        city:
          type: string
        number:
          type: string
        postalCode:
          type: string
        country:
          type: string
    ContactInformation:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int32
        description:
          type: string
        value:
          type: string